fix: iabtcf doesn't run tests correctly
