--- .mocharc.json
@@ -1,12 +1,0 @@
-{
-  "require": [
-    "ts-node/register",
-    "jsdom-global/register",
-    "source-map-support/register"
-  ],
-  "node-option": [
-    "experimental-specifier-resolution=node",
-    "loader=ts-node/esm"
-  ],
-  "spec": "./modules/**/test/Root.test.ts"
-}
