+++ modules/cmpapi/.mocharc.json
@@ -2,11 +2,12 @@
   "require": [
     "ts-node/register",
     "jsdom-global/register",
-    "source-map-support/register"
+    "source-map-support/register",
+    "test/hook.mjs"
   ],
   "node-option": [
     "experimental-specifier-resolution=node",
     "loader=ts-node/esm"
   ],
-  "spec": "./test/Root.test.ts"
+  "spec": "./test/**/*.test.ts"
 }
