+++ modules/cmpapi/global.d.ts
@@ -1,0 +1,11 @@
+// eslint-disable-next-line max-len
+/* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-unused-vars, no-unused-vars, @typescript-eslint/no-var-requires */
+interface Window {
+    __tcfapi: (
+        command: any,
+        version: any,
+        callback: (...params: any) => void,
+        ...param: any
+    ) => void;
+
+}
