+++ modules/core/test/GVL.test.ts
@@ -7,11 +7,13 @@ import {XMLHttpTestTools} from '@iabtechlabtcf/testing';
 import {Json} from '../src/Json';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const vendorlistJson = require('@iabtechlabtcf/testing/lib/vendorlist/v2/vendor-list-v24.json');
+import vendorListJson from '../../testing/lib/mjs/vendorlist/v2/vendor-list-v24.json';
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const translationJson = require('@iabtechlabtcf/testing/lib/vendorlist/v2/purposes-fr.json');
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const vendorlistJson22 = require('@iabtechlabtcf/testing/lib/vendorlist/v2.2/vendor-list.json');
+import translationJson from '../../testing/lib/mjs/vendorlist/v2/purposes-fr.json';
+import {VersionOrVendorList} from '../lib/mjs';
+import vendorlistJson22 from '../../testing/lib/vendorlist/v2.2/vendor-list.json');
+
+const vendorlistJson: any = vendorListJson as unknown as VersionOrVendorList;
 
 describe('GVL', (): void => {
 
@@ -129,7 +131,7 @@ describe('GVL', (): void => {
 
   it('should clone all values', (): void => {
 
-    const gvl: GVL = new GVL(vendorlistJson);
+    const gvl: GVL = new GVL(vendorlistJson as unknown as VersionOrVendorList);
     const clone: GVL = gvl.clone();
 
     assertPopulated(gvl);
